full workflow to tag a bioconductor package and selected representative content with EDAM terms

full workflow to tag a bioconductor package and selected representative content with EDAM terms

Usage

tag_bioc(
  pkgname = "tximeta",
  url =
    "https://bioconductor.org/packages/release/bioc/vignettes/tximeta/inst/doc/tximeta.html"
)

Arguments

pkgname

character(1)

url

URL for representative content, in html or pdf

Value

a data.frame with biocViews and EDAM suggestions

Examples

if (interactive() & nchar(Sys.getenv("OPENAI_API_KEY")>0)) {
ti = tag_bioc()
bs = tag_bioc(pkg="Biostrings", 
  url="https://bioconductor.org/packages/release/bioc/vignettes/Biostrings/inst/doc/Biostrings2Classes.pdf")
library(DT)
ndf = rbind(ti, bs)
datatable(ndf)
}
#> Using model = "gpt-4.1".
#> Using model = "gpt-4.1".
#> 1/0 packages newly attached/loaded, see sessionInfo() for details.